Don't do backpressure release logic when not called for
We currently don't check to see if we are throttled when calling
release backpressure. This means that we execute all the release
backpressure code even if we aren't throttled. This can happen in
several different ways resulting in spurious logic being run and lots of
noise in the logs.

Currently on OSX where we aren't using KQueue one-shot each message sent
can result in "backpressure release" happening.