Have Machida build during CI
This commit adds a very basic Makefile for Machida that will cause it to
be run as part our standard CI and fail the build if it can no longer
build. There's a stub placeholder for executing tests.

This currently doesn't work when run from within the `machida` directory
due to an unknown bug that Dipin and I have to persue. Given the value
in running Machida builds as part of the CI, we are commiting this as
is.

Closes #807