Remove spdlog dependency and C++ buffer classes
spdlog was not providing much of an advantage over using standard
C/C++ printing techniques (cout or printf) and added an unnecessary
dependency to the C++ API. Removing it makes it easier to set up CI
builds.

The C++ buffer classes were not being used anywhere. Developing a
general buffering solution for application developers is outside of
the scope of this library at this time.