Get C++ code to build during CI
With this change, changes that break C++ examples and the C++ API should
in theory be caught by CI. We will now:

* Build the C++ API
* Build the C++ examples from the book
* Build the MarketSpread C++ application

As part of this change, I reorganized where the API code lives.
Previously, it was in lib/wallaroo which at the time made some sense.
However, looking at it now, the CPP API is a user of that Wallaroo API
rather than part of it. Therefore, Andy and I decided to it make more
sense to move the `cpp_api` code to the top level of the Walllaroo
repository.

All C++ example apps have Makefiles that in theory should be able to be
run from the top-level Wallaroo directory or from with their given
directory. However, because the Makefiles depend on the C++ API code
having been built, you can not actually do something like:

```bash
cd book/examples/cpp/counter-app
make test
```

because the Makefile rule for building the C++ API isn't known at that
level. It's only know if you do something like:

make build-book-examples-cpp-counter-app-all

from the top level Wallaroo directory as the Makefile that gets created
for there knows about all the build rules. This is something we need to
look at addressing.

Also note, I discovered as part of this that the cmake build for the C++
API is somewhat broken in that the `install/local` command that is
supposed to only "install" within the build directory attempts to
install to the greater filesystem as well. Why? Not sure. There's some
deep ugly cmake magic there.

Its my general opinion after going through this that at this time we do
not need cmake. It's an extra dependency that we are carrying around. We
should replace it, time permitting, with some standard Makefiles,
perhaps a configure script and be done with it. If we need to solve the
problems that cmake looks to solve, we can evaluate options in the
future.

I found a couple issues that need to be address but not as part of this
PR. These have been opened as issues #834 and #835.