Temporarily turn off building Machida with resilience on
Currently, it will segfault if the programmer doesn't supply serialize
and deserialize functions. We are going to do a release with resilience
off and then follow on shortly thereafter with another release that will
work and...

print a message if serialize or deserialize isnt available to say that
we are running in a degraded mode without resilience.