Fix issue where state computations in Python API could not filter
Filtering was broken in the Python API because of the change from
returning the message to returning a tuple of the message and a
boolean indicating whether or not state shouldb persisted. Originally,
filtering was done by returning `None` from the state computation. But
now there are two values, and the first one can be
`None`. Consequently the output was never seen as being `None`, so
messages were never filtered.