Add default serialization for Machida applications
If the application developer does not provide serialization functions
then default serialization functions are used. This lets us build
Machida with resilience turned on by default and have programs run
even if they don't provide their own serialization functions. It also
means that developers do not have to provide their own serialization
functions for multiworker applications if the default ones are
sufficient for their needs.

The default serialization functions are in the `wallaroo` module. If
Machida does not find serialization functions in the main application
module then it accesses the `wallaroo` module through the main
application module and then accesses the serialization functions from
there. They use the built-in `pickle` package.

Since Machida applications can now work with resilience even if they
don't provide serialization functions, this commit turns on resilience
by default in the Machida makefile.

This system assumes that the `wallaroo` module was imported from the main
application module and is publicly available from that module. If the
programmer has done something to violate this assumption then this
system WILL NOT WORK.