Hash _BoundaryId fields rather than `this`
Due to using `(digestof this).hash()` instead of hashing the fields
of _BoundaryId (a string and a U128), attempts to retrieve an existing
DataReceiver weren't always successful.
The new hash method for _BoundaryId XORs the hashes of its fields (following
the pattern in net/http/_host_service.pony).

Additionally:
- the window increment test is embedded in the sequence-window application
and can be enabled to run in real-time and Fail the process if the test fails
by using `--define/-D validate`.

close #832