Don't "corrupt" tcp connection state machines
Each of our classes that handle a TCPConnection had a couple subtle race
conditions in them caused by sloppy handling of state machine variables.
The final fix for this is to move state for each of these classes into a
proper state machine so everything can't get wildly out of whack.

An example issue is #869 where we could continue trying to send data
after our connection had been dropped. In order to fix, we need to:

Not update readable and writeable to true unless we have a connection.
Hard close in _pending_read when we hit an error (close() can leave us
in an invalid state).
Check to make sure that _shutdown_peer isn't set when writing (or
already did this for reading).

Closes #869