Add information about object serialization in Python
The new text gives more detail about what what the user needs to do in
order to serialize objects. Some objects (such as third-party objects
that wrap C data) may not be able to be serialized with Python's
`pickle` module, so the user should be aware of what they are
serializing and make sure that their serialization mechanism supports
it.

Fixes #860

[skip ci]