Add test coverage for DataChannel
This is not complete coverage but mirrors what TCPConnection in the Pony
standard library has.

This requires some changes to work. In particular, I removed unneeded
dependencies from DataReceiver and DataReceivers. Additionally, stating
up a real MetricsSink resulted in a Timer being created in order to run
the reconnect protocol.

MetricsSink is now an interface and the concrete type we previously had
is now ReconnectingMetricsSink

Closes #636