Don't FAIL application when TCPSource gets a "bad" message
Previously, if you sent a non-decodable message to a `TCPSource`, it
would call `Fail()` and stop the entire application. This is great for
debugging but not so good for regular production usage.

Now, we will only `Fail()` when in debug mode. This allows us to catch
errors more easily during development (assuming you are running in
debug) while avoiding shutting down a production application.

Closes #510