Use "Epoch friendly" timestamps in Giles
Giles currently uses unadjusted monotonic timestamps that can only be
compared to one another. With this change, timestamps are nanoseconds
since the unix epoch allowing it to be compared to times from other
applications like say, a Python wallaroo application.

This would allow for comparing of a timestamp from inside a C++ or
Python application with Giles timestamps and track the amount of time
spent at different points not just in a Wallaroo application but
throughout its lifecycle.