Fork "PonyTest" into lib/Sendence
There are few areas where PonyTest falls short of what we currently
need. Amongst these:

The means by which "long tests" are scheduled can fail in constrained
resource environments like TravisCI. The issue is addressable but design
flaws in PonyTest need to be fixed.

Additionally addative functionality would be very useful to us, for
example, adding a `reject_action` to go along with `expect_action` which
would fail a test if a given action was taken.

After speaking with Joe, a Pony core team member, he expressed wanting
to send changes through the Pony RFC process. I don't think this is a
good plan for us because:

1) A solid test harness is of very high value to Sendence. Being able to
move quickly to add new features is a big win for us.
2) I doubt that PonyTest will be receiving many updates meaning that we
aren't likely to miss out on valuable features.

As we move forward, I think that our test harness will become more and
more important to us and we need one that fits our needs. The Pony
community is small enough that it is unlikely that anyone will be
stepping forward with improvements anytime soon for PonyTest.

This PR move PonyTest into lib/sendence and renames it Connemara. All
usage of PonyTest in our repo is removed as part of this commit and
replaced by usage of Connemara. Once this is merged, I'll start work on
improvements to Connemara.

Ideally, Connemara will remain backwards compatible with PonyTest so
that any test that PonyTest can run, Connemara can run. However, I
expect we will eventually reach a point where we decide to forgo that
compatibility in the name of features that we need.

Once this is merged to master, I'll start work on fixing PonyTest issues
and adding additional functionality to Connemara.

Why Connemara?

1) I expect this to exist outside of Wallaroo eventually so a "unique"
name seems to make sense
2) The Connemara Pony is one of my favorite kinds of Ponies.

https://en.wikipedia.org/wiki/Connemara_pony