Add documentation for when state change object cannot be found
This change adds documentation for what happens when the a call to
`w_state_change_repository_lookup_by_name` cannot find the specified
state change object, which will happen if it is given a string that
does not match the name of an existing state change object. Currently
this will return a pointer to the Pony `None` object. This function
will should only be called inside the `compute` method of a stateful
computation, which receives a pointer to the Pony `None` object as the
parameter `none`, so the user can check to make sure they have a valid
state change object by comparing the returned pointer to `none`.

We may want to change the API at some point to return `null` rather
than `None`, which would probably be more in keeping with what a C++
programmer would expect.

Fixes issue #965.

[skip ci]