Fix recovery bug
Fail() on data channel failing to listen uncovered a
latent bug in initializer recovery where it still created
a ClusterInitializer that thought the topology was yet
to be initialized.  This meant that when another worker
later recovered, the initializer would try to create a
new data listener in response to the recovering worker
identifying its data port. Since we were already listening
on the initializer data port, the initializer would
fail to listen a second time and Fail().  The ClusterInitializer
is now aware of the difference between initial spin up
and recovery.