Don't wait to start counting reconnected boundaries
Prior to this change, the RecoveryReplay protocol
required that boundaries can only reconnect
after we receive all boundary counts for workers.
But this is not actually necessary. We can keep
track of reconnecting boundaries at the same
time as we wait for boundary counts. If all
boundaries have reconnected by the time we get
the last boundary count, we can start replaying
right away. This fixes a bug where reconnection
attempts led to Fail() during recovery.