Add parallel state computations to initialization
When an app developer uses the new
`to_parallel` method while defining an application,
we mark that computation as part of a stateless
partition. During initialization, we coalesce
that computation with surrounding stateless
computations, calculate the number of steps
needed for the partition, and add all those
steps to the init graph for each worker according
to its share.