Add the ability to specify TCP source host and port in topology setup
The host and service are now specified as part of the topology setup
code rather than being passed as arguments on the command line.

This is a step toward providing a general mechanism for specifying
sources other than raw TCP listeners (such as Kafka sources).

The celsius.pony example has been updated to use this new mechanism.