Specify source listener rather than always using TCP source listener
The source listener builder builder is now part of the
SourceInformation object that is passed to the `from` method when
creating a pipeline. This work lays the groundwork for eventually
being able to specify other kinds of source listener builder builders
so that we can support other types of sources (like Kafka).