Create a mostly generic SourceSpecification interface
There is now a SourceSpecification interface that can be used to
create concrete classes that specify information about a
source. Currenlty only TCPSourceInformation implements this interface.

The SourceBuilder class, which is used to set up the source, would
ideally be generic across source types, but it currenty has an apply
method that return a TCPSourceNotify object. I need to figure out how
to handle this so that either the apply method can be generic or so
that it is no longer part of the interface ... or maybe something
else.