Add primitives for parsing TCP source and sink config from CLI
These primitives take the command line arguments, parse them, and
return objects that represent TCP source and sink configurations. The
application developer is now responsible for writing code that
configures sinks and sources, so these primitives are provided to make
it easy for developers to read these options from the command line as
long as the assumed format is suffient for their needs.