Make source configuration a required argument to `new_pipeline`
The `Application.new_pipeline(...)` method now requires a source
config objec to configure the pipeline's source. This replaces the old
system where the source config was passes after creating the new
pipeline, and if nothing was passed then a TCP source was assumed. The
old behavior wasn't something that a user would expect; it makes more
sense to require the source configuration information.