Update C++ API to use explict TCP sources and sinks
The C++ API was updated to use the new source and sink APIs. The
changes were made at the API level, there were no changes made to
applications. This means that currently the applications can only use
TCP sources. At some later point we will have to expand the API so
that other types of sources can be specified.