Add StatelessPartitionRouter and populate it
The step/s with edges into a stateless partition
will use a StatelessPartitionRouter. This is
populated during local initialization. We first
wait for all steps in the partition to be created
and added to our map from step ids to steps. We
then use that information plus the information about
distribution across workers in PreStatelessData
to populate this router.