Change step sequence id generator to support 1 to many (#1050)
With this change, how steps generate sequence ids now supports 1 to
many.

Previously, we would generate a single sequence id at the time the
step's run method was call and never again. In a 1 to many scenario this
would result in every outgoing message having the same id.

This does not complete 1 to many support but is a step along the way.
The next step after this is to implement 1 to many and update how
outgoing messages are tracked to incoming in order to support 1 to many.