Support 1 to many messaging
This commit is the final commit needed to add 1 to many to Wallaroo.
With this commit, a Wallaroo developer can return an Array[Out] to
produce multiple outgoing messages for a single incoming message.

There are two temporary apps in 'testing/correctness/apps' that can be
used to sanity check that the basic functionality works.

- `one_to_many` to verify that stateless computations work correctly.
- `one_to_many_stateless` to verify that stateful computations work
correctly.