Generic source sink (#1048)
This change makes sources and sinks generic, and implements TCP-specific sources and sinks based on these generic sources and sink interfaces. This work makes it possible to add new kinds of sources and sinks, including ones for Kafka.

The Pony applications have been updated to use the new APIs.

The C++ API, Python API, and Atkin have all been modified to assume that they are using TCP sources and sinks. This allows the applications to run unmodified, but it also means that there will need to be additional work on the APIs in order to allow other types of sources and sinks to be used by these systems.