Routing for multiworker stateless partition
We use the seq id of a message to enact
stateless round robin routing across a
stateless partition. Each stateless partition
has an internal set of sequential partition
ids from 0 to (size of partition - 1). We
mod the seq id against the size f partition
to determine where the next message is to
be routed, either to a local step or to
a ProxyRouter if the target is on another
worker.