Add one-to-many testing to sequence window
Adds logic to sequence window correctness testing to do both 1-to-1 and
1-to-many testing. I chose to incorporate one to many messaging (and
filtering) into the sequence window rather than doing a new sequence
window so we would have fewer tests to run and less code to maintain. I
don't believe there is any real downside to having the logic in the
existing code. If there is a downside, we can always make the:

          .to[U64]({(): MaybeOneToMany => MaybeOneToMany})

be conditionally compiled in.