Support tracking of "fractional" messages (#1093)
In order to correctly do deduplication of messages, we need to be able to know they are unique. We have part of this with out unique message id that is assigned at the time a message enters wallaroo via a source. With the introduction of "1 to many", this was no longer sufficient. One message with id "10" could because five with that id making deduplication impossible.

This commit adds full support for the concept of "fractional ids" that was first discussed in the fall of 2016 when we were working on our watermarking and exactly-once ideas.