Get rid of init files
Init files (the ones that could be passed as part of pipeline
creation, that would feed an initial batch of data to the pipeline)
were a hack that we don't really need. What's more, Wallaroo use the
source addresses (`--in` argument) from the command line to determine
where to send the contents of these files; this is a problem because
now applications are supposed to configure their own sources and the
Wallaroo application framework is supposed to be unaware of them.

The input addresses were being passed around the configuration code in
order to support init files; now that they have been removed the input
addresses are no longer passed around in Wallaroo.

The Wactor system still depends on input addresses for the command
line, so there is a new config parser method specifically for the
Wactor system that gets them.