Remove option to turn coalescing off
Because we needed it for tests, we added the option to turn off coalescing. We've long since outgrown the need to turn off coalescing for tests. However, we've continued to support it as part of the API. This has lead to numerous bugs and other problems. 

We are removing this option from the API because we don't want to support it once we go public.

This commit removes coalescing code as well as supporting tests.

Closes #1109 
Closes #1071