Add support for general sources and sinks in the Python API
The Python will now support general sources and sinks in a
pipeline. The application developer must supply a configuration object
for the type of source and sink, and they must write code that
determines what options to use for those configurations. For example,
if a TCP source is used then the developer must do something like
hardcode the host and port or parse them from the command line
arguments.

Classes are provided for TCP sources and sinks. Also, two helper
functions have been provided to parse the host/port information from
the command line if the application developer uses our standard `--in`
and `--out` arguments.