Add Python support for Kafka sources and sinks
Kafka sources and sinks are now supported in Python applications. This
allows Python application to use these sinks and sources in addition
to the exiting TCP sinks and sources.

The Python API was updated so that decoders are passed as arguments to
the source configation objects and encoders are passed as arguments to
the sink configuration objects. This reflects the way the Pony API is
arranged, and also results in less casting of different kinds of
objects.

In order to make it easier to create the `KafkaConfig` objects for the
sink and source, the Pony API was modified and the creation of the the
config objects was moved into factory classes. This made it easier to
use independent of the Pony CLI parser classes.

The documentation has been updated to reflect the API changes.

There is an example called `celsius-kafka` in `book/examples/python`.

Closes #1025