Create a package for "kill this program" exit conditions
We currently have wallaroo/fail. This creates a corresponding class in
sendence/mort that can we can refactor to using. I'm not doing that now
so that we don't create a mess prior to open sourcing, but with this
change, we can open an issue to make the change.

Also adds an `Unreachable` that will also exit but will be more obvious
in code that it's unreachable. This way we no longer need code like:

```pony
// unreachable
Fail()
```

instead we can replace it with the more concise

```pony
Unreachable()
```