Fix race condition in sending watermarks to event log
This race condition could allow our watermarks that we have written to
the log to go backwards. This could happen because:

1- our watermarking algo allows proposed watermarks to go backwards
2- we handle this in the Acker but not sending a proposed watermark to
the event log if its less than what we already flushed

However the "what we already flushed" is incorrect. That value was what
had already been written to the event log. We don't want to try to flush
a watermark if we already sent one to be flushed that was a higher
value.

This change implements a fix for the aforementioned race condition.