Update book "getting started" to be Python specific
This change update the book's "Getting Started" section to be
Python-specific. Since Python the only supported API at this point is
Python it makes sense to immediately expose the user to a Python
app. Introducing a Pony app and then not talking about Pony apps after
that seems confusing and also delays giving the user information about
how they should work with Wallaroo.

The setup instructions for Linux and Mac OS were updated to include
instructions for building Machida.

The Python `intro.md` and `building.md` were updated to reflect the
fact that Machida should have already been built by the time the user
reaches this section, and `building.md` was renamed to
`running-a-wallaroo-python-application.md` to better reflect the
content of the file.

The `examples/python/celsius/celsius.py` application was updated to
match the text in `book/getting-started/run-a-application.md`.

Fixes #1165