Quiet down "reconnecting" related metrics sink messages
For our initial release, we aren't telling people to run a metrics receiver in
examples. This would result in a lot of noise in STDOUT. Until metrics are
optional, see: https://github.com/WallarooLabs/wallaroo/issues/766
we are turning of printing that we are reconnecting. Once metrics are optional,
we can revisit this and decide when and how often we should print.