Move "installation" instructions for Giles
Per issue #1287, Andy feels this is a better placement. The MacOS
instructions were in a different order than the Linux ones. This
different order would result in the existing Machida install
instructions to not work. As part of this PR, I reworked to order of the
MacOS instructions to match the Linux ones which should allow everything
to correct work as all required directories will be in place before
building machida and giles.

[skip ci]

Closes #1287