Improve Python example READMEs (#1246)
Update Python examples to "Andy format".

Andy's description and reasoning:

"There's a section that describes the application, including the input
and output formats. This gives the reader an idea of what the program
does and how it does it.

The run section specifies how many shells the user will need to run
the example, and clearly describes what should be done in each
shell. There is information about what directory commands should be
run from.

I think it is reasonable to assume that some people will encounter the
Wallaroo repo before reading the book, and one of the first things
they will do is look at the examples. These changes make the examples
more of a useful (if not ideal) starting point for learning about
Wallaroo."