Provide a more "external user" friendly README (#1317)
This is based on feedback from the entire team. There is additional
content to come. I wanted to get a version in now that is complete for
any information that is available.

The documentation section will be fleshed out once the new website is
online.
There will be a status section for features once we've generated that
content.
There will be a short description of each feature once we have generated
that content.

[skip ci]