Add link to support opens when people go to open an issue (#1350)
SUPPORT.md is a "magic" GitHub file. By including this file in the repo,
a link to "support" will appear when a user goes to open an issue. With
this addition, we will have

- SUPPORT.md
- CONTRIBUTING.md
- CODE_OF_CONDUCT.md

which will result in:

Before you open an issue please review the support, contributing, and
conduct guidelines for this repository.

Being displayed on the new issues page.

[skip ci]