Updates message count sent in for Run A Application Example
The amount of data sent in to the Celsius example needed
to be updated in order to get continuous metrics updates in the UI.
This was done in order for the user to be able to navigate the UI and
be able to spend some time looking at it without everything coming to a
halt.

Adds note to user that the data file is being repeated.

Depends on #1355 and #1356 before this can be merged

Closes #1334

[skip ci]