Move HubProtocol to MonHub
This group of changes moves the HubProtocol to the MonHub from
phoenix_tcp. This chage was necessary in order for phoenix_tcp to be
a useable tcp transport for phoenix channels  by anyone instead of
just the MonitoringHub. The changes are as follows:

Update mix.lock to point to phoenix_tcp master

This updates the branch and commit we’re locked on, to no longer rely
on the
HubProtocol being in phoenix_tcp.

Add phoenix dependency to MonitoringHubUtils

Needed in order for the TCPSerializer to adopt the
Phoenix.Transports.Serializer behaviour.

Adds generic TCPServer to MonitoringHubUtils

Adds a generic TCPServer to MonitoringHubUtils. This server does not
send back
any data to Wallaroo due to Wallaroo doing nothing with responses at
the moment.

Adds TCPSerializer to MonitoringHubUtils

Adds a TCPSerializer to MonitoringHubUtils which implements the
Phoenix.Transports.Serializer behaviour. This is needed in order to
encode and
decode messages using the HubProtocol to Phoenix.Socket.Message structs.

Adds HubProtocol to MonitoringHubUtils

Adds HubProtocol to MonitoringHubUtils. This is used to decode the
various
messages received via Wallaroo.

Updates MetricsSocket transport options for TCP messages

Adds the MonitoringHubUtils TCPServer and TCPSerializer as options for
the TCP
Transport for the MetricsSocket. This allows them to be used when
receiving
messages to the Metrics Channel via TCP.

Updates dev + prod config options for MetricsReporterUI

Adds tcp_server and protocol options to the tcp transport options.
This allows them to be used as defaults until the the appropriate
socket can
determine which server and serializer to use.