Add a book section about long running data processing etc
This adds a section addressing "Wallaroo and Long-running data
processing and other workflows". This is meant to address concerns
about whether or not wallaroo is just a stream processing system. Vid
requested this section in order to address the question: "What about
batch processing?" which has come up in meetings with potential
customers.

Closes #1503

[skip ci]