Use variable for wallaroo version where needed in docs
It was becoming very error prone to update in all the locations that
were needed. Doing global search and replace was bound to cause
problems.

As an example, I went to change `0.1.0` to `release-0.1.1` and picked up
our pony-kafka version in bundle.json files. Even if I was to constrain
to the book/ directory, eventually a mistake would be made.