Remove print statements from Python market spread
Prior to this commit the Python version of market spread printed
messages to the console when orders were rejected, and if trades on a
certain symbol should be rejected. This could result in lots of
messages being printed, which could in turn cause performance problems
and make it difficult to see important messages coming from Wallaroo.

The `print` statements have been removed from the application.

Closes #1380