Get rid of `cover.docx`
The `cover.docx` file was being generated from the Google doc that
represents the book cover. It isn't used directly to generate the
`cover.jpg` file, so it doesn't serve any purpose in the repo. It
mostly confuses people about the actual process of building
`cover.jpg`, which is described in issue #1534.

This commit removes the file from the repository, and hopefully
reduces the amount of confusion around an already somewhat confusing
process.

Closes #1534

[skip ci]