Remove `cover.pdf` from the repository
The `cover.pdf` file is an intermediary file for generating the cover
of the book. It doesn't really make sense to keep it in the repository
because it is generated from a google doc and then used to generate
the jpg that is used by gitbook. Keeping it in the repository only
causes confusion.

Fixes #1498

[skip ci]