Fix bug in submakefile include logic
Prior to this submakefile includes had an edge case where they
didn't work correctly if the same name was used in another part
of the directory tree due to how the includes were handled
individually. This commit resolves it so that the includes are
now all handled as a group ensuring correctness and consistency.

resolves #1231