Document signing the CLA in `CONTRIBUTING.md`
Contributors to Wallaroo are required to sign a contributor license
agreement (CLA) before their pull requests can be accepted. We use the
CLA Assistant (https://cla-assistant.io) to handle this. This commit
adds a note to the "Pull request" section that has a link that the
user can follow to sign the agreement, another link to just the text
of the agreement, and a not saying that if they don't sign it ahead of
time then they will be prompted to sign it when they submit a PR. This
should make the process clear to the contributor.

Closes #1578

[skip ci]