Make `Makefile` control variables explicit instead of implicit
Prior to this commit a bad setting or an error in one Makefile
could lead to other rules defined in other Makefiles not working
at all. This commit changes the control variables used by `rules.mk`
to be explicit (must be either `true` or `false`) and to be unique
for every makefile so they're not shared between makefiles anymore.

resolves #1587