Have `make test` run `make unit-tests` and `make integration-tests`
Prior to this commit we only had a `make test` target (and related
for all subdirectories) targets. This commit adds the
`make unit-tests` and `make integration-tests` targets as required
to split up running of unit tests from integration tests.

Additinally this commit renames the related control variable to be
`UNIT_TEST_COMMAND`.

This commit also re-categorizes any custom test targets to be
dependencies of either `make unit-tests` or `make integration-tests`
as appropriate.

resolves #1130