Update README to remove references to most GitBook builds
We've turned off building gitbook from anything other than the release
branch. Building gitbook from master was creating an odd number of
clones in the GitHub stats. We want to track clones to see how many
people are trying out Wallaroo (or as an indicator). All the GitBook
cloning that results in ¯\_(ツ)_/¯ number of "unique clones" in GitHub
was making that difficult.

[skip ci]