Undo `val` to `iso` changes for `recovery`
Previously, as part of commit 2feeb40, we switched from sending
`Array[ByteSeq] val` to `Array[ByteSeq] iso` to work around a `val`
tracing bug. As part of the update to the latest ponylang ponyc,
that workaround is no longer required and this commit changes
the `Array[ByteSeq] iso` back to `Array[ByteSeq] val`.