Remove mention of generating a data file in Running an Application
The "Running a Wallaroo Application" section of the book was changed
to use a pre-generated data file, but there was still a piece of text
that talked about using a data file that was generated earlier. This
has been changed to say that the example will use pre-generated
data. This should prevent confusion on the user's part.

Fixes #1577

[skip ci]