Drop metrics messages when not connected
Prior to this change, we infinitely queued metrics messages if we
weren't connected to a metrics receiving source. This would cause large
amounts of never ending memory growth. Our previous strategy was to shed
load, at some point that got accidentally changed to "queue if you've
never connected" and "shed load if at some point connected but not now".

This change causes metrics messages to always be shed if not connected.