Add basic Wallaroo reverse app that uses Go
This Wallaroo application is written in Pony but the logic for the
decoder, encoder, and reverse function are implemented in the Go.

Go places constraints on how Go pointers can be used outside of Go
code. Consequently the Go code returns handles to Wallaroo rather than
pointers to objects themselves. The constraints are described here:

https://golang.org/cmd/cgo/#hdr-Passing_pointers

This is a very rough version of the reverse application that ignores
all kinds of important issues. For example, any object that is
returned from Go to Wallaroo will never be garbage collected because
the Go map that maps handles to objects never releases it.