Initial commit of Go API experimental work
This commit includes a rough Wallaroo API for Go, a word count example
that uses this API, and a very rough word reverse application that
does not use the API.

Go places constraints on how Go pointers can be used outside of Go
code. Consequently the Go code returns handles to Wallaroo rather than
pointers to objects themselves. The constraints are described here:

https://golang.org/cmd/cgo/#hdr-Passing_pointers

The API is designed around this constraint. Instead of passing
pointers to Go objects out to Pony, the Go objects are assigned IDs
and stored in a map, and hthen the IDs are returned to Pony. Pony
stores these IDs and uses them when calling Go functions.